Add files via upload
In many React applications, you’ll send the data to an external service, like
a Web API. When the service resolves, you’ll often show a success
message, redirect the user, or do both.
To simulate an API, add a setTimeout function in the handleSubmit
function. This will create an asynchronous operation that waits a certain
amount of time before completing, which behaves similarly to a request for
external data. Then use the useState Hook to create a submitting variable
export default App;
and a setSubmitting function. Call setSubmitting(true) when the data is
submitted and call setSubmitting(false) when the timeout is resolved:

Now you have a basic form that handles the submit event inside the React
component. You’ve connected it to your JSX using the onSubmit event
handler and you are using Hooks to conditionally display an alert while the
handleSubmit event is running