Add files via upload
— Collecting Form Data Using Uncontrolled
Components

Form using uncontrolled component

In this step, you’ll collect form data using uncontrolled components. An
uncontrolled component is a component that does not have a value set by
React. Instead of setting the data on the component, you’ll connect to the on
Change event to collect the user input. As you build the components, you’ll
learn how React handles different input types and how to create a reusable
function to collect form data into a single object.
By the end of this step, you’ll be able to build a form using different form
elements, including dropdowns and checkboxes. You’ll also be able to
collect, submit, and display form data.
Note: In most cases, you’ll use controlled components for your React
application. But it’s a good idea to start with uncontrolled components
so that you can avoid subtle bugs or accidental loops that you might
introduce when incorrectly setting a value.
Currently, you have a form that can submit information, but there is nothing
to submit. The form has a single <input> element, but you are not
collecting or storing the data anywhere in the component. In order to be
able to store and process the data when the user submits a form, you’ll need
to create a way to manage state. You’ll then need to connect to each input
using an event handler.
Inside App.js , use the useReducer Hook to create a formData object and a
setFormData function. For the reducer function, pull the name and value
from the event.target object and update the state by spreading the
current state while adding the name and value at the end. This will create a
state object that preserves the current state while overwriting specific values
as they change: